Add argument check to bash script, rm pwsh script as well

+++ initialize_app.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ $# -eq 0 ]; then
+    echo "No directory name supplied"
+    echo "Usage: bash initialize_app.sh dirname"
+    exit -1
+fi
+
 # Recreate folder
 rm -rf ../$1
 mkdir ../$1
@@ -8,8 +14,9 @@ mkdir ../$1
 shopt -s extglob
 cp -a !(node_modules) ../$1
 
-# Remove install script
+# Remove install scripts
 rm -f ../$1/initialize_app.sh
+rm -f ../$1/initialize_app.ps1
 
 # Initialize a git repository and create a first commit
 cd ../$1
